65576::【后台】【部门协作】监管状态变更中咨询业务线增加接口判断未完成订单

+++ ehealth-recipe-core/src/main/java/recipe/regulation/RegulationRecipeQueryService.java
@@ -41,7 +41,7 @@ public class RegulationRecipeQueryService implements IRegulationRecipeQueryServi
     @Override
     public Integer findUnfinishedRecipe(Integer organId) {
         logger.info("findUnfinishedRecipe organId={}", organId);
-        Integer size = recipeDAO.getUnfinishedRecipe(organId, PayConstant.PAY_FLAG_NOT_PAY);
-        return size;
+        Long size = recipeDAO.getUnfinishedRecipe(organId, PayConstant.PAY_FLAG_NOT_PAY);
+        return size.intValue();
     }
 }
